% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{add_tags}
\alias{add_tags}
\title{Adds or overwrites one or more tags for the specified Amazon SageMaker resource}
\usage{
add_tags(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{The Amazon Resource Name (ARN) of the resource that you want to tag.}

\item{Tags}{An array of \code{Tag} objects. Each tag is a key-value pair. Only the \code{key} parameter is required. If you don't specify a value, Amazon SageMaker sets the value to an empty string.}
}
\description{
Adds or overwrites one or more tags for the specified Amazon SageMaker resource. You can add tags to notebook instances, training jobs, models, endpoint configurations, and endpoints.
}
\details{
Each tag consists of a key and an optional value. Tag keys must be unique per resource. For more information about tags, see \href{http://docs.aws.amazon.com/awsaccountbilling/latest/aboutv2/cost-alloc-tags.html#allocation-what}{Using Cost Allocation Tags} in the \emph{AWS Billing and Cost Management User Guide}.
}
\examples{

}
