% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{create_code_repository}
\alias{create_code_repository}
\title{Create a git repository as a resource in your Amazon SageMaker account}
\usage{
create_code_repository(
  CodeRepositoryName = "string",
  GitConfig = list(
    RepositoryUrl = "string",
    Branch = "string",
    SecretArn = "string"
  )
)
}
\arguments{
\item{CodeRepositoryName}{[required] The name of the git repository. The name must have 1 to 63 characters. Valid characters are a-z, A-Z, 0-9, and - (hyphen).}

\item{GitConfig}{[required] Specifies details about the repository, including the URL where the repository is located, the default branch, and credentials to use to access the repository.}
}
\description{
Create a git repository as a resource in your Amazon SageMaker account. You can associate the repository with notebook instances so that you can use git source control for the notebooks you create. The git repository is a resource in your Amazon SageMaker account, so it can be associated with more than one notebook instance, and it persists independently from the lifecycle of any notebook instances it is associated with.
}
\details{
The repository can be hosted either in \href{http://docs.aws.amazon.com/codecommit/latest/userguide/welcome.html}{AWS CodeCommit} or in any other git repository.
}
\examples{

}
