% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{create_endpoint_config}
\alias{create_endpoint_config}
\title{Creates an endpoint configuration that Amazon SageMaker hosting services uses to deploy models}
\usage{
create_endpoint_config(EndpointConfigName, ProductionVariants, Tags = NULL,
  KmsKeyId = NULL)
}
\arguments{
\item{EndpointConfigName}{The name of the endpoint configuration. You specify this name in a \href{http://docs.aws.amazon.com/sagemaker/latest/dg/API_CreateEndpoint.html}{CreateEndpoint} request.}

\item{ProductionVariants}{An array of \code{ProductionVariant} objects, one for each model that you want to host at this endpoint.}

\item{Tags}{An array of key-value pairs. For more information, see \href{http://docs.aws.amazon.com/awsaccountbilling/latest/aboutv2/cost-alloc-tags.html#allocation-what}{Using Cost Allocation Tags} in the \emph{AWS Billing and Cost Management User Guide}.}

\item{KmsKeyId}{The Amazon Resource Name (ARN) of a AWS Key Management Service key that Amazon SageMaker uses to encrypt data on the storage volume attached to the ML compute instance that hosts the endpoint.}
}
\description{
Creates an endpoint configuration that Amazon SageMaker hosting services uses to deploy models. In the configuration, you identify one or more models, created using the \code{CreateModel} API, to deploy and the resources that you want Amazon SageMaker to provision. Then you call the \href{http://docs.aws.amazon.com/sagemaker/latest/dg/API_CreateEndpoint.html}{CreateEndpoint} API.
}
\details{
Use this API only if you want to use Amazon SageMaker hosting services to deploy models into production.

In the request, you define one or more \code{ProductionVariant}s, each of which identifies a model. Each \code{ProductionVariant} parameter also describes the resources that you want Amazon SageMaker to provision. This includes the number and type of ML compute instances to deploy.

If you are hosting multiple models, you also assign a \code{VariantWeight} to specify how much traffic you want to allocate to each model. For example, suppose that you want to host two models, A and B, and you assign traffic weight 2 for model A and 1 for model B. Amazon SageMaker distributes two-thirds of the traffic to Model A, and one-third to model B.
}
\examples{

}
