% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{create_notebook_instance}
\alias{create_notebook_instance}
\title{Creates an Amazon SageMaker notebook instance}
\usage{
create_notebook_instance(NotebookInstanceName, InstanceType, RoleArn,
  SubnetId = NULL, SecurityGroupIds = NULL, KmsKeyId = NULL,
  Tags = NULL, LifecycleConfigName = NULL, DirectInternetAccess = NULL)
}
\arguments{
\item{NotebookInstanceName}{The name of the new notebook instance.}

\item{InstanceType}{The type of ML compute instance to launch for the notebook instance.}

\item{RoleArn}{When you send any requests to AWS resources from the notebook instance, Amazon SageMaker assumes this role to perform tasks on your behalf. You must grant this role necessary permissions so Amazon SageMaker can perform these tasks. The policy must allow the Amazon SageMaker service principal (sagemaker.amazonaws.com) permissions to assume this role. For more information, see \href{http://docs.aws.amazon.com/sagemaker/latest/dg/sagemaker-roles.html}{Amazon SageMaker Roles}.

To be able to pass this role to Amazon SageMaker, the caller of this API must have the \code{iam:PassRole} permission.}

\item{SubnetId}{The ID of the subnet in a VPC to which you would like to have a connectivity from your ML compute instance.}

\item{SecurityGroupIds}{The VPC security group IDs, in the form sg-xxxxxxxx. The security groups must be for the same VPC as specified in the subnet.}

\item{KmsKeyId}{If you provide a AWS KMS key ID, Amazon SageMaker uses it to encrypt data at rest on the ML storage volume that is attached to your notebook instance.}

\item{Tags}{A list of tags to associate with the notebook instance. You can add tags later by using the \code{CreateTags} API.}

\item{LifecycleConfigName}{The name of a lifecycle configuration to associate with the notebook instance. For information about lifestyle configurations, see notebook-lifecycle-config.}

\item{DirectInternetAccess}{Sets whether Amazon SageMaker provides internet access to the notebook instance. If you set this to \code{Disabled} this notebook instance will be able to access resources only in your VPC, and will not be able to connect to Amazon SageMaker training and endpoint services unless your configure a NAT Gateway in your VPC.

For more information, see appendix-notebook-and-internet-access. You can set the value of this parameter to \code{Disabled} only if you set a value for the \code{SubnetId} parameter.}
}
\description{
Creates an Amazon SageMaker notebook instance. A notebook instance is a machine learning (ML) compute instance running on a Jupyter notebook.
}
\details{
In a \code{CreateNotebookInstance} request, specify the type of ML compute instance that you want to run. Amazon SageMaker launches the instance, installs common libraries that you can use to explore datasets for model training, and attaches an ML storage volume to the notebook instance.

Amazon SageMaker also provides a set of example notebooks. Each notebook demonstrates how to use Amazon SageMaker with a specific algorithm or with a machine learning framework.

After receiving the request, Amazon SageMaker does the following:
\enumerate{
\item Creates a network interface in the Amazon SageMaker VPC.
\item (Option) If you specified \code{SubnetId}, Amazon SageMaker creates a network interface in your own VPC, which is inferred from the subnet ID that you provide in the input. When creating this network interface, Amazon SageMaker attaches the security group that you specified in the request to the network interface that it creates in your VPC.
\item Launches an EC2 instance of the type specified in the request in the Amazon SageMaker VPC. If you specified \code{SubnetId} of your VPC, Amazon SageMaker specifies both network interfaces when launching this instance. This enables inbound traffic from your own VPC to the notebook instance, assuming that the security groups allow it.
}

After creating the notebook instance, Amazon SageMaker returns its Amazon Resource Name (ARN).

After Amazon SageMaker creates the notebook instance, you can connect to the Jupyter server and work in Jupyter notebooks. For example, you can write code to explore a dataset that you can use for model training, train a model, host models by creating Amazon SageMaker endpoints, and validate hosted models.

For more information, see \href{http://docs.aws.amazon.com/sagemaker/latest/dg/how-it-works.html}{How It Works}.
}
\examples{

}
