% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{create_presigned_notebook_instance_url}
\alias{create_presigned_notebook_instance_url}
\title{Returns a URL that you can use to connect to the Jupyter server from a notebook instance}
\usage{
create_presigned_notebook_instance_url(NotebookInstanceName,
  SessionExpirationDurationInSeconds = NULL)
}
\arguments{
\item{NotebookInstanceName}{The name of the notebook instance.}

\item{SessionExpirationDurationInSeconds}{The duration of the session, in seconds. The default is 12 hours.}
}
\description{
Returns a URL that you can use to connect to the Jupyter server from a notebook instance. In the Amazon SageMaker console, when you choose \code{Open} next to a notebook instance, Amazon SageMaker opens a new tab showing the Jupyter server home page from the notebook instance. The console uses this API to get the URL and show the page.
}
\details{
You can restrict access to this API and to the URL that it returns to a list of IP addresses that you specify. To restrict access, attach an IAM policy that denies access to this API unless the call comes from an IP address in the specified list to every AWS Identity and Access Management user, group, or role used to access the notebook instance. Use the \code{NotIpAddress} condition operator and the \code{aws:SourceIP} condition context key to specify the list of IP addresses that you want to have access to the notebook instance. For more information, see nbi-ip-filter.
}
\examples{

}
