% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{create_workteam}
\alias{create_workteam}
\title{Creates a new work team for labeling your data}
\usage{
create_workteam(
  WorkteamName = "string",
  MemberDefinitions = list(
    list(
      CognitoMemberDefinition = list(
        UserPool = "string",
        UserGroup = "string",
        ClientId = "string"
      )
    )
  ),
  Description = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{WorkteamName}{[required] The name of the work team. Use this name to identify the work team.}

\item{MemberDefinitions}{[required] A list of \code{MemberDefinition} objects that contains objects that identify the Amazon Cognito user pool that makes up the work team. For more information, see \href{http://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-identity-pools.html}{Amazon Cognito User Pools}.

All of the \code{CognitoMemberDefinition} objects that make up the member definition must have the same \code{ClientId} and \code{UserPool} values.}

\item{Description}{[required] A description of the work team.}

\item{Tags}{}
}
\description{
Creates a new work team for labeling your data. A work team is defined by one or more Amazon Cognito user pools. You must first create the user pools before you can create a work team.
}
\details{
You cannot create more than 25 work teams in an account and region.
}
\examples{

}
