% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{delete_algorithm}
\alias{delete_algorithm}
\title{Removes the specified algorithm from your account}
\usage{
delete_algorithm(
  AlgorithmName = "string"
)
}
\arguments{
\item{AlgorithmName}{[required] The name of the algorithm to delete.}
}
\description{
Removes the specified algorithm from your account.
}
\examples{

}
