% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{delete_code_repository}
\alias{delete_code_repository}
\title{Deletes the specified Git repository from your account}
\usage{
delete_code_repository(CodeRepositoryName)
}
\arguments{
\item{CodeRepositoryName}{[required] The name of the Git repository to delete.}
}
\description{
Deletes the specified Git repository from your account.
}
\section{Accepted Parameters}{
\preformatted{delete_code_repository(
  CodeRepositoryName = "string"
)
}
}

