% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{delete_endpoint}
\alias{delete_endpoint}
\title{Deletes an endpoint}
\usage{
delete_endpoint(EndpointName)
}
\arguments{
\item{EndpointName}{The name of the endpoint that you want to delete.}
}
\description{
Deletes an endpoint. Amazon SageMaker frees up all of the resources that were deployed when the endpoint was created.
}
\details{
Amazon SageMaker retires any custom KMS key grants associated with the endpoint, meaning you don't need to use the \href{http://docs.aws.amazon.com/kms/latest/APIReference/API_RevokeGrant.html}{RevokeGrant} API call.
}
\examples{

}
