% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{delete_endpoint_config}
\alias{delete_endpoint_config}
\title{Deletes an endpoint configuration}
\usage{
delete_endpoint_config(EndpointConfigName)
}
\arguments{
\item{EndpointConfigName}{The name of the endpoint configuration that you want to delete.}
}
\description{
Deletes an endpoint configuration. The \code{DeleteEndpointConfig} API deletes only the specified configuration. It does not delete endpoints created using the configuration.
}
\examples{

}
