% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{delete_model}
\alias{delete_model}
\title{Deletes a model}
\usage{
delete_model(ModelName)
}
\arguments{
\item{ModelName}{The name of the model to delete.}
}
\description{
Deletes a model. The \code{DeleteModel} API deletes only the model entry that was created in Amazon SageMaker when you called the \href{http://docs.aws.amazon.com/sagemaker/latest/dg/API_CreateModel.html}{CreateModel} API. It does not delete model artifacts, inference code, or the IAM role that you specified when creating the model.
}
\examples{

}
