% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{delete_notebook_instance}
\alias{delete_notebook_instance}
\title{Deletes an Amazon SageMaker notebook instance}
\usage{
delete_notebook_instance(NotebookInstanceName)
}
\arguments{
\item{NotebookInstanceName}{The name of the Amazon SageMaker notebook instance to delete.}
}
\description{
Deletes an Amazon SageMaker notebook instance. Before you can delete a notebook instance, you must call the \code{StopNotebookInstance} API.
}
\details{
When you delete a notebook instance, you lose all of your data. Amazon SageMaker removes the ML compute instance, and deletes the ML storage volume and the network interface associated with the notebook instance.
}
\examples{

}
