% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{delete_notebook_instance_lifecycle_config}
\alias{delete_notebook_instance_lifecycle_config}
\title{Deletes a notebook instance lifecycle configuration}
\usage{

  delete_notebook_instance_lifecycle_config(NotebookInstanceLifecycleConfigName)
}
\arguments{
\item{NotebookInstanceLifecycleConfigName}{[required] The name of the lifecycle configuration to delete.}
}
\description{
Deletes a notebook instance lifecycle configuration.
}
\section{Accepted Parameters}{
\preformatted{delete_notebook_instance_lifecycle_config(
  NotebookInstanceLifecycleConfigName = "string"
)
}
}

