% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{delete_tags}
\alias{delete_tags}
\title{Deletes the specified tags from an Amazon SageMaker resource}
\usage{
delete_tags(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{The Amazon Resource Name (ARN) of the resource whose tags you want to delete.}

\item{TagKeys}{An array or one or more tag keys to delete.}
}
\description{
Deletes the specified tags from an Amazon SageMaker resource.
}
\details{
To list a resource's tags, use the \code{ListTags} API.
}
\examples{

}
