% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_algorithm}
\alias{describe_algorithm}
\title{Returns a description of the specified algorithm that is in your account}
\usage{
describe_algorithm(
  AlgorithmName = "string"
)
}
\arguments{
\item{AlgorithmName}{[required] The name of the algorithm to describe.}
}
\description{
Returns a description of the specified algorithm that is in your account.
}
\examples{

}
