% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_code_repository}
\alias{describe_code_repository}
\title{Gets details about the specified git repository}
\usage{
describe_code_repository(
  CodeRepositoryName = "string"
)
}
\arguments{
\item{CodeRepositoryName}{[required] The name of the git repository to describe.}
}
\description{
Gets details about the specified git repository.
}
\examples{

}
