% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_compilation_job}
\alias{describe_compilation_job}
\title{Returns information about a model compilation job}
\usage{
describe_compilation_job(CompilationJobName)
}
\arguments{
\item{CompilationJobName}{[required] The name of the model compilation job that you want information about.}
}
\description{
Returns information about a model compilation job.
}
\details{
To create a model compilation job, use CreateCompilationJob. To get information about multiple model compilation jobs, use ListCompilationJobs.
}
\section{Accepted Parameters}{
\preformatted{describe_compilation_job(
  CompilationJobName = "string"
)
}
}

