% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_endpoint_config}
\alias{describe_endpoint_config}
\title{Returns the description of an endpoint configuration created using the \code{CreateEndpointConfig} API}
\usage{
describe_endpoint_config(
  EndpointConfigName = "string"
)
}
\arguments{
\item{EndpointConfigName}{[required] The name of the endpoint configuration.}
}
\description{
Returns the description of an endpoint configuration created using the \code{CreateEndpointConfig} API.
}
\examples{

}
