% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_hyper_parameter_tuning_job}
\alias{describe_hyper_parameter_tuning_job}
\title{Gets a description of a hyperparameter tuning job}
\usage{
describe_hyper_parameter_tuning_job(
  HyperParameterTuningJobName = "string"
)
}
\arguments{
\item{HyperParameterTuningJobName}{[required] The name of the tuning job to describe.}
}
\description{
Gets a description of a hyperparameter tuning job.
}
\examples{

}
