% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_labeling_job}
\alias{describe_labeling_job}
\title{Gets information about a labeling job}
\usage{
describe_labeling_job(LabelingJobName)
}
\arguments{
\item{LabelingJobName}{[required] The name of the labeling job to return information for.}
}
\description{
Gets information about a labeling job.
}
\section{Accepted Parameters}{
\preformatted{describe_labeling_job(
  LabelingJobName = "string"
)
}
}

