% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_notebook_instance}
\alias{describe_notebook_instance}
\title{Returns information about a notebook instance}
\usage{
describe_notebook_instance(NotebookInstanceName)
}
\arguments{
\item{NotebookInstanceName}{[required] The name of the notebook instance that you want information about.}
}
\description{
Returns information about a notebook instance.
}
\section{Accepted Parameters}{
\preformatted{describe_notebook_instance(
  NotebookInstanceName = "string"
)
}
}

