% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_notebook_instance_lifecycle_config}
\alias{describe_notebook_instance_lifecycle_config}
\title{Returns a description of a notebook instance lifecycle configuration}
\usage{

  describe_notebook_instance_lifecycle_config(NotebookInstanceLifecycleConfigName)
}
\arguments{
\item{NotebookInstanceLifecycleConfigName}{[required] The name of the lifecycle configuration to describe.}
}
\description{
Returns a description of a notebook instance lifecycle configuration.
}
\details{
For information about notebook instance lifestyle configurations, see \href{http://docs.aws.amazon.com/sagemaker/latest/dg/notebook-lifecycle-config.html}{Step 2.1: (Optional) Customize a Notebook Instance}.
}
\section{Accepted Parameters}{
\preformatted{describe_notebook_instance_lifecycle_config(
  NotebookInstanceLifecycleConfigName = "string"
)
}
}

