% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_subscribed_workteam}
\alias{describe_subscribed_workteam}
\title{Gets information about a work team provided by a vendor}
\usage{
describe_subscribed_workteam(WorkteamArn)
}
\arguments{
\item{WorkteamArn}{[required] The Amazon Resource Name (ARN) of the subscribed work team to describe.}
}
\description{
Gets information about a work team provided by a vendor. It returns details about the subscription with a vendor in the AWS Marketplace.
}
\section{Accepted Parameters}{
\preformatted{describe_subscribed_workteam(
  WorkteamArn = "string"
)
}
}

