% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_training_job}
\alias{describe_training_job}
\title{Returns information about a training job}
\usage{
describe_training_job(TrainingJobName)
}
\arguments{
\item{TrainingJobName}{[required] The name of the training job.}
}
\description{
Returns information about a training job.
}
\section{Accepted Parameters}{
\preformatted{describe_training_job(
  TrainingJobName = "string"
)
}
}

