% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_transform_job}
\alias{describe_transform_job}
\title{Returns information about a transform job}
\usage{
describe_transform_job(
  TransformJobName = "string"
)
}
\arguments{
\item{TransformJobName}{[required] The name of the transform job that you want to view details of.}
}
\description{
Returns information about a transform job.
}
\examples{

}
