% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_workteam}
\alias{describe_workteam}
\title{Gets information about a specific work team}
\usage{
describe_workteam(
  WorkteamName = "string"
)
}
\arguments{
\item{WorkteamName}{[required] The name of the work team to return a description of.}
}
\description{
Gets information about a specific work team. You can see information such as the create date, the last updated date, membership information, and the work team's Amazon Resource Name (ARN).
}
\examples{

}
