% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{list_compilation_jobs}
\alias{list_compilation_jobs}
\title{Lists model compilation jobs that satisfy various filters}
\usage{
list_compilation_jobs(
  NextToken = "string",
  MaxResults = 123,
  CreationTimeAfter = as.POSIXct("2015-01-01"),
  CreationTimeBefore = as.POSIXct("2015-01-01"),
  LastModifiedTimeAfter = as.POSIXct("2015-01-01"),
  LastModifiedTimeBefore = as.POSIXct("2015-01-01"),
  NameContains = "string",
  StatusEquals = "INPROGRESS"|"COMPLETED"|"FAILED"|"STARTING"|"STOPPING"|"STOPPED",
  SortBy = "Name"|"CreationTime"|"Status",
  SortOrder = "Ascending"|"Descending"
)
}
\arguments{
\item{NextToken}{If the result of the previous \code{ListCompilationJobs} request was truncated, the response includes a \code{NextToken}. To retrieve the next set of model compilation jobs, use the token in the next request.}

\item{MaxResults}{The maximum number of model compilation jobs to return in the response.}

\item{CreationTimeAfter}{A filter that returns the model compilation jobs that were created after a specified time.}

\item{CreationTimeBefore}{A filter that returns the model compilation jobs that were created before a specified time.}

\item{LastModifiedTimeAfter}{A filter that returns the model compilation jobs that were modified after a specified time.}

\item{LastModifiedTimeBefore}{A filter that returns the model compilation jobs that were modified before a specified time.}

\item{NameContains}{A filter that returns the model compilation jobs whose name contains a specified string.}

\item{StatusEquals}{A filter that retrieves model compilation jobs with a specific DescribeCompilationJobResponse\$CompilationJobStatus status.}

\item{SortBy}{The field by which to sort results. The default is \code{CreationTime}.}

\item{SortOrder}{The sort order for results. The default is \code{Ascending}.}
}
\description{
Lists model compilation jobs that satisfy various filters.
}
\details{
To create a model compilation job, use CreateCompilationJob. To get information about a particular model compilation job you have created, use DescribeCompilationJob.
}
\examples{

}
