% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{list_endpoints}
\alias{list_endpoints}
\title{Lists endpoints}
\usage{
list_endpoints(SortBy = NULL, SortOrder = NULL, NextToken = NULL,
  MaxResults = NULL, NameContains = NULL, CreationTimeBefore = NULL,
  CreationTimeAfter = NULL, LastModifiedTimeBefore = NULL,
  LastModifiedTimeAfter = NULL, StatusEquals = NULL)
}
\arguments{
\item{SortBy}{Sorts the list of results. The default is \code{CreationTime}.}

\item{SortOrder}{The sort order for results. The default is \code{Ascending}.}

\item{NextToken}{If the result of a \code{ListEndpoints} request was truncated, the response includes a \code{NextToken}. To retrieve the next set of endpoints, use the token in the next request.}

\item{MaxResults}{The maximum number of endpoints to return in the response.}

\item{NameContains}{A string in endpoint names. This filter returns only endpoints whose name contains the specified string.}

\item{CreationTimeBefore}{A filter that returns only endpoints that were created before the specified time (timestamp).}

\item{CreationTimeAfter}{A filter that returns only endpoints that were created after the specified time (timestamp).}

\item{LastModifiedTimeBefore}{A filter that returns only endpoints that were modified before the specified timestamp.}

\item{LastModifiedTimeAfter}{A filter that returns only endpoints that were modified after the specified timestamp.}

\item{StatusEquals}{A filter that returns only endpoints with the specified status.}
}
\description{
Lists endpoints.
}
\examples{

}
