% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{list_labeling_jobs_for_workteam}
\alias{list_labeling_jobs_for_workteam}
\title{Gets a list of labeling jobs assigned to a specified work team}
\usage{
list_labeling_jobs_for_workteam(
  WorkteamArn = "string",
  MaxResults = 123,
  NextToken = "string",
  CreationTimeAfter = as.POSIXct("2015-01-01"),
  CreationTimeBefore = as.POSIXct("2015-01-01"),
  JobReferenceCodeContains = "string",
  SortBy = "CreationTime",
  SortOrder = "Ascending"|"Descending"
)
}
\arguments{
\item{WorkteamArn}{[required] The Amazon Resource Name (ARN) of the work team for which you want to see labeling jobs for.}

\item{MaxResults}{The maximum number of labeling jobs to return in each page of the response.}

\item{NextToken}{If the result of the previous \code{ListLabelingJobsForWorkteam} request was truncated, the response includes a \code{NextToken}. To retrieve the next set of labeling jobs, use the token in the next request.}

\item{CreationTimeAfter}{A filter that returns only labeling jobs created after the specified time (timestamp).}

\item{CreationTimeBefore}{A filter that returns only labeling jobs created before the specified time (timestamp).}

\item{JobReferenceCodeContains}{A filter the limits jobs to only the ones whose job reference code contains the specified string.}

\item{SortBy}{The field to sort results by. The default is \code{CreationTime}.}

\item{SortOrder}{The sort order for results. The default is \code{Ascending}.}
}
\description{
Gets a list of labeling jobs assigned to a specified work team.
}
\examples{

}
