% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{list_subscribed_workteams}
\alias{list_subscribed_workteams}
\title{Gets a list of the work teams that you are subscribed to in the AWS Marketplace}
\usage{
list_subscribed_workteams(NameContains = NULL, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{NameContains}{A string in the work team name. This filter returns only work teams whose name contains the specified string.}

\item{NextToken}{If the result of the previous \code{ListSubscribedWorkteams} request was truncated, the response includes a \code{NextToken}. To retrieve the next set of labeling jobs, use the token in the next request.}

\item{MaxResults}{The maximum number of work teams to return in each page of the response.}
}
\description{
Gets a list of the work teams that you are subscribed to in the AWS Marketplace. The list may be empty if no work team satisfies the filter specified in the \code{NameContains} parameter.
}
\section{Accepted Parameters}{
\preformatted{list_subscribed_workteams(
  NameContains = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

