% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{list_tags}
\alias{list_tags}
\title{Returns the tags for the specified Amazon SageMaker resource}
\usage{
list_tags(ResourceArn, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{ResourceArn}{The Amazon Resource Name (ARN) of the resource whose tags you want to retrieve.}

\item{NextToken}{If the response to the previous \code{ListTags} request is truncated, Amazon SageMaker returns this token. To retrieve the next set of tags, use it in the subsequent request.}

\item{MaxResults}{Maximum number of tags to return.}
}
\description{
Returns the tags for the specified Amazon SageMaker resource.
}
\examples{

}
