% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_package.R
\docType{package}
\name{paws.sagemaker-package}
\alias{paws.sagemaker}
\alias{paws.sagemaker-package}
\title{paws.sagemaker: Amazon SageMaker Service}
\description{
Definition of the public APIs exposed by SageMaker
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
