% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{stop_training_job}
\alias{stop_training_job}
\title{Stops a training job}
\usage{
stop_training_job(TrainingJobName)
}
\arguments{
\item{TrainingJobName}{The name of the training job to stop.}
}
\description{
Stops a training job. To stop a job, Amazon SageMaker sends the algorithm the \code{SIGTERM} signal, which delays job termination for 120 seconds. Algorithms might use this 120-second window to save the model artifacts, so the results of the training is not lost.
}
\details{
Training algorithms provided by Amazon SageMaker save the intermediate results of a model training job. This intermediate data is a valid model artifact. You can use the model artifacts that are saved when Amazon SageMaker stops a training job to create a model.

When it receives a \code{StopTrainingJob} request, Amazon SageMaker changes the status of the job to \code{Stopping}. After Amazon SageMaker stops the job, it sets the status to \code{Stopped}.
}
\examples{

}
