% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{stop_transform_job}
\alias{stop_transform_job}
\title{Stops a transform job}
\usage{
stop_transform_job(
  TransformJobName = "string"
)
}
\arguments{
\item{TransformJobName}{[required] The name of the transform job to stop.}
}
\description{
Stops a transform job.
}
\details{
When Amazon SageMaker receives a \code{StopTransformJob} request, the status of the job changes to \code{Stopping}. After Amazon SageMaker stops the job, the status is set to \code{Stopped}. When you stop a transform job before it is completed, Amazon SageMaker doesn't store the job's output in Amazon S3.
}
\examples{

}
