% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{update_code_repository}
\alias{update_code_repository}
\title{Updates the specified git repository with the specified values}
\usage{
update_code_repository(
  CodeRepositoryName = "string",
  GitConfig = list(
    SecretArn = "string"
  )
)
}
\arguments{
\item{CodeRepositoryName}{[required] The name of the git repository to update.}
}
\description{
Updates the specified git repository with the specified values.
}
\examples{

}
