% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{delete_code_repository}
\alias{delete_code_repository}
\title{Deletes the specified git repository from your account}
\usage{
delete_code_repository(
  CodeRepositoryName = "string"
)
}
\arguments{
\item{CodeRepositoryName}{[required] The name of the git repository to delete.}
}
\description{
Deletes the specified git repository from your account.
}
\examples{

}
