% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_endpoint}
\alias{describe_endpoint}
\title{Returns the description of an endpoint}
\usage{
describe_endpoint(
  EndpointName = "string"
)
}
\arguments{
\item{EndpointName}{[required] The name of the endpoint.}
}
\description{
Returns the description of an endpoint.
}
\examples{

}
