% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_model}
\alias{describe_model}
\title{Describes a model that you created using the CreateModel API}
\usage{
describe_model(ModelName)
}
\arguments{
\item{ModelName}{[required] The name of the model.}
}
\description{
Describes a model that you created using the \code{CreateModel} API.
}
\section{Accepted Parameters}{
\preformatted{describe_model(
  ModelName = "string"
)
}
}

