% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.sagemaker_operations.R
\name{describe_model}
\alias{describe_model}
\title{Describes a model that you created using the \code{CreateModel} API}
\usage{
describe_model(ModelName)
}
\arguments{
\item{ModelName}{The name of the model.}
}
\description{
Describes a model that you created using the \code{CreateModel} API.
}
\examples{

}
